refactor(server): consolidate duplicated oauth guards into shared leaf module
Adds the guards.ts module referenced by jwt.ts in the previous commit and
removes the copy-pasted isRecord/isLocalhost/invalidToken definitions from
internal.ts, supabase.ts, and keycloak.ts.
